requirements.txt
requests
aiohttp
PyYAML
pytest